requirements.txt
discord.py
python-dotenv
requests
httpx
fastapi
uvicorn
beautifulsoup4